fix: cast power terms before CockroachDB random arithmetic

+++ src/plans.ts
@@ -1092,8 +1092,8 @@ function failJobs (schema: string, table: string, where: string, output: string)
              ELSE now() + LEAST(
                retry_delay_max::float8,
                retry_delay::float8 * (
-                2 ^ LEAST(16, retry_count + 1) / 2 +
-                2 ^ LEAST(16, retry_count + 1) / 2 * random()
+                (2 ^ LEAST(16, retry_count + 1) / 2)::float8 +
+                (2 ^ LEAST(16, retry_count + 1) / 2)::float8 * random()
                )
              ) * interval '1s'
         END as start_after,
